Move scheduler functionality to a super module
