+++ src/rdd/rdd.rs
@@ -12,9 +12,9 @@ use crate::dependency::Dependency;
 use crate::error::{Error, Result};
 use crate::partial::CountEvaluator;
 use crate::partitioner::{HashPartitioner, Partitioner};
+use crate::scheduler::TaskContext;
 use crate::serializable_traits::{AnyData, Data, Func, SerFunc};
 use crate::split::Split;
-use crate::task::TaskContext;
 use crate::utils::random::{BernoulliSampler, PoissonSampler, RandomSampler};
 use crate::{utils, Fn, SerArc, SerBox};
 use fasthash::MetroHasher;
