+++ src/scheduler/base_scheduler.rs
@@ -4,35 +4,17 @@ use std::sync::Arc;
 use std::thread;
 use std::time::{Duration, Instant};
 
-use crate::dag_scheduler::{CompletionEvent, FetchFailedVals};
 use crate::dependency::{Dependency, ShuffleDependencyTrait};
 use crate::env;
 use crate::error::{Error, Result};
-use crate::job::JobTracker;
 use crate::rdd::{Rdd, RddBase};
-use crate::result_task::ResultTask;
+use crate::scheduler::{
+    CompletionEvent, FetchFailedVals, JobTracker, ResultTask, Stage, TaskBase, TaskContext,
+    TaskOption,
+};
 use crate::serializable_traits::{Data, SerFunc};
 use crate::shuffle::ShuffleMapTask;
-use crate::stage::Stage;
-use crate::task::{TaskBase, TaskContext, TaskOption};
 use dashmap::DashMap;
-
-pub trait Scheduler {
-    fn start(&self);
-    fn wait_for_register(&self);
-    fn run_job<T: Data, U: Data, F>(
-        &self,
-        rdd: &dyn Rdd<Item = T>,
-        func: F,
-        partitions: Vec<i64>,
-        allow_local: bool,
-    ) -> Vec<U>
-    where
-        Self: Sized,
-        F: Fn(Box<dyn Iterator<Item = T>>) -> U;
-    fn stop(&self);
-    fn default_parallelism(&self) -> i64;
-}
 
 pub(crate) type EventQueue = Arc<DashMap<usize, VecDeque<CompletionEvent>>>;
 
@@ -547,27 +529,19 @@ macro_rules! impl_common_scheduler_funcs {
             self.shuffle_to_map_stage.get(&id).unwrap().clone()
         }
 
-
         fn unregister_map_output(&self, shuffle_id: usize, map_id: usize, server_uri: String) {
-            self.map_output_tracker.unregister_map_output(
-                shuffle_id,
-                map_id,
-                server_uri
-            )
+            self.map_output_tracker
+                .unregister_map_output(shuffle_id, map_id, server_uri)
         }
 
         fn register_shuffle(&self, shuffle_id: usize, num_maps: usize) {
-            self.map_output_tracker.register_shuffle(
-                shuffle_id,
-                num_maps
-            )
+            self.map_output_tracker
+                .register_shuffle(shuffle_id, num_maps)
         }
 
         fn register_map_outputs(&self, shuffle_id: usize, locs: Vec<Option<String>>) {
-            self.map_output_tracker.register_map_outputs(
-                shuffle_id,
-                locs
-            )
+            self.map_output_tracker
+                .register_map_outputs(shuffle_id, locs)
         }
 
         fn remove_output_loc_from_stage(&self, shuffle_id: usize, map_id: usize, server_uri: &str) {
@@ -602,6 +576,5 @@ macro_rules! impl_common_scheduler_funcs {
         fn get_next_task_id(&self) -> usize {
             self.next_task_id.fetch_add(1, Ordering::SeqCst)
         }
-
     };
 }
