+++ src/scheduler/result_task.rs
@@ -5,8 +5,8 @@ use std::sync::Arc;
 
 use crate::env;
 use crate::rdd::Rdd;
+use crate::scheduler::{Task, TaskBase, TaskContext};
 use crate::serializable_traits::Data;
-use crate::task::{Task, TaskBase, TaskContext};
 use crate::SerBox;
 use serde_derive::{Deserialize, Serialize};
 use serde_traitobject::{Deserialize, Serialize};
